Move Series.factorize to IndexOpsMixin.
